User's ownership of cartItem check was added, and some others fixes

+++ src/main/java/com/samoilenko/onlinebookstore/service/impl/ShoppingCartServiceImpl.java
@@ -14,9 +14,10 @@ import com.samoilenko.onlinebookstore.repository.BookRepository;
 import com.samoilenko.onlinebookstore.repository.CartItemRepository;
 import com.samoilenko.onlinebookstore.repository.ShoppingCartRepository;
 import com.samoilenko.onlinebookstore.service.ShoppingCartService;
-import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
@@ -28,7 +29,7 @@ public class ShoppingCartServiceImpl implements ShoppingCartService {
     private final CartItemRepository cartItemRepository;
 
     @Override
-    @Transactional
+    @Transactional(readOnly = true)
     public ShoppingCartResponseDto getCartByUserId(Long id) {
         return shoppingCartMapper.toResponseDto(
                 shoppingCartRepository.findCartByUserId(id)
@@ -36,6 +37,7 @@ public class ShoppingCartServiceImpl implements ShoppingCartService {
     }
 
     @Override
+    @Transactional
     public CartItemResponseDto addCartItem(Long userId, CartItemRequestDto requestDto) {
         CartItem cartItem = cartItemMapper.toEntity(requestDto);
         cartItem.setBook(bookRepository.findById(requestDto.bookId()).orElseThrow(
@@ -53,24 +55,34 @@ public class ShoppingCartServiceImpl implements ShoppingCartService {
     }
 
     @Override
-    public CartItemResponseDto update(Long cartItemId, UpdateRequestDto requestDto) {
-        CartItem itemToUpdate = cartItemRepository.findById(cartItemId).orElseThrow(
-                () -> new EntityNotFoundException("Cart item with id " + cartItemId + " not found")
-        );
+    @Transactional
+    public CartItemResponseDto update(Long userId, Long cartItemId, UpdateRequestDto requestDto) {
+        CartItem itemToUpdate = getCartItem(cartItemId);
+        checkUserOwnershipOfItem(userId, itemToUpdate);
         itemToUpdate.setQuantity(requestDto.quantity());
         return cartItemMapper.toResponseDto(cartItemRepository.save(itemToUpdate));
     }
 
     @Override
-    public void deleteCartItemById(Long cartItemId) {
-        validateId(cartItemId);
+    @Transactional
+    public void deleteCartItemById(Long userId, Long cartItemId) {
+        CartItem cartItemToDelete = getCartItem(cartItemId);
+        checkUserOwnershipOfItem(userId, cartItemToDelete);
         cartItemRepository.deleteById(cartItemId);
     }
 
-    private void validateId(Long id) {
-        if (!cartItemRepository.existsById(id)) {
-            throw new EntityNotFoundException("Cart item with id " + id + " not found");
-        }
+    private CartItem getCartItem(Long cartItemId) {
+        return cartItemRepository.findById(cartItemId).orElseThrow(
+                () -> new EntityNotFoundException("Cart item with id " + cartItemId + " not found")
+        );
+    }
 
+    private void checkUserOwnershipOfItem(Long userId, CartItem cartItem) {
+        Long shoppingCartId = shoppingCartRepository.findCartByUserId(userId).getId();
+        if (shoppingCartId != cartItem.getShoppingCart().getId()) {
+            throw new AccessDeniedException(
+                    "User does not have permission to modify this cart item"
+            );
+        }
     }
 }
